requirements.txt
Django==2.2.19
django-cors-headers==3.11.0
pillow==9.4.0
djoser==2.1.0
django-filter==21.1
psycopg2-binary==2.8.6
python-dotenv==0.21.1